(DataView) Moves dnd context up above table
