+++ src/components/DataView/DataViewTBody/DataViewTBodyTypes.ts
@@ -1,4 +1,3 @@
-import { ReactNode } from "react";
 import { MosaicObject } from "@root/types";
 import { DataViewProps } from "../DataViewTypes";
 import { DataViewDisplayListProps } from "../DataViewDisplayList";
@@ -16,5 +15,3 @@ export interface DataViewTBodyProps {
 	columns: DataViewProps["columns"];
 	hasActions: boolean;
 }
-
-export type DataViewTBodySortableProps = Pick<DataViewTBodyProps, "data" | "transformedData" | "onReorder"> & { children: ReactNode }
